+++ frontend/src/supabase/repositories/index.ts
@@ -22,9 +22,9 @@ export * from './base'
 
 // Domain repositories
 export * from './activities'
+export * from './registrations'
 
 // TODO: Add remaining repositories
-// export * from './registrations'
 // export * from './users'
 // export * from './balances'
 // export * from './sections'